+++ app/src/main/kotlin/org/fossify/calendar/fragments/DayTimelineFragmentsHolder.kt
@@ -1,0 +1,7 @@
+package org.fossify.calendar.fragments
+
+import org.fossify.calendar.helpers.DAILY_VIEW
+
+class DayTimelineFragmentsHolder : DayFragmentsHolder() {
+    override val viewType = DAILY_VIEW
+}