+++ app/src/main/kotlin/org/fossify/calendar/helpers/MyWidgetMonthlyProvider.kt
@@ -63,7 +63,7 @@ class MyWidgetMonthlyProvider : AppWidgetProvider() {
     private fun setupDayOpenIntent(context: Context, views: RemoteViews, id: Int, dayCode: String) {
         (context.getLaunchIntent() ?: Intent(context, SplashActivity::class.java)).apply {
             putExtra(DAY_CODE, dayCode)
-            putExtra(VIEW_TO_OPEN, DAILY_VIEW)
+            putExtra(VIEW_TO_OPEN, AGENDA_VIEW)
             val pendingIntent = PendingIntent.getActivity(context, Integer.parseInt(dayCode), this, PendingIntent.FLAG_IMMUTABLE)
             views.setOnClickPendingIntent(id, pendingIntent)
         }