+++ app/src/main/kotlin/org/fossify/calendar/interfaces/DayPage.kt
@@ -1,0 +1,9 @@
+package org.fossify.calendar.interfaces
+
+interface DayPage {
+    var mListener: NavigationListener?
+
+    fun updateCalendar()
+
+    fun printCurrentView()
+}